Pass event label as tag instead of field
So InfluxDB doesn't support GROUP BYs on fields. If we want to GROUP BY
something, it needs to be sent as a tag. This, however, might be a problem
in the future because if there's a lot of random stuff in tags, that will
increase the series cardinality a lot which is not good for performance:

https://docs.influxdata.com/influxdb/v1.7/concepts/glossary/#series-cardinality

Because InfluxDB requires some field to be sent along with the tags, we
now send 1 as the value for these view events. Maybe we can pretend it
represents a single view.