[Backend] Update TypeScript settings
Target was updated from ES6 to ES2021 as backend runs on Node 20. Also lib is specified to be ES2023 so that we are able to use methods like `Array.at()` or `Array.toSorted()` with types. Also strict is changed to true but all the individual rules that currently cause problems have been disabled.