[Class statistics] Fix filters on Credits gained tab not working properly
Previously the filters used the Credits earned filter, but that's not correct as it uses number of all credits, no matter when they were attained. Previously the Personal study plan filter didn't have an effect on the table either even though the table header suggested it did. Now HOPS credits are used in the table if Personal study plan filter is active.